move credentials logic off command.Command
the credential stuff isn't tied to the docker client. This moves logic to helper functions that both clients can use
