+++ pkg/docker/docker_command.go
@@ -80,29 +80,10 @@ func (c *DockerCommand) Push(ctx context.Context, image string) error {
 	return c.exec(ctx, nil, nil, nil, "", []string{"push", image})
 }
 
+// TODO[md]: this doesn't need to be on the interface, move to auth handler
 func (c *DockerCommand) LoadUserInformation(ctx context.Context, registryHost string) (*command.UserInfo, error) {
 	console.Debugf("=== DockerCommand.LoadUserInformation %s", registryHost)
 
-	conf := config.LoadDefaultConfigFile(os.Stderr)
-	credsStore := conf.CredentialsStore
-	if credsStore == "" {
-		authConf, err := loadAuthFromConfig(conf, registryHost)
-		if err != nil {
-			return nil, err
-		}
-		return &command.UserInfo{
-			Token:    authConf.Password,
-			Username: authConf.Username,
-		}, nil
-	}
-	credsHelper, err := loadAuthFromCredentialsStore(ctx, credsStore, registryHost)
-	if err != nil {
-		return nil, err
-	}
-	return &command.UserInfo{
-		Token:    credsHelper.Secret,
-		Username: credsHelper.Username,
-	}, nil
 }
 
 func (c *DockerCommand) CreateTarFile(ctx context.Context, image string, tmpDir string, tarFile string, folder string) (string, error) {
@@ -151,6 +132,7 @@ func (c *DockerCommand) CreateAptTarFile(ctx context.Context, tmpDir string, apt
 	}
 
 	return aptTarFile, nil
+	return loadUserInformation(ctx, registryHost)
 }
 
 func (c *DockerCommand) Inspect(ctx context.Context, ref string) (*image.InspectResponse, error) {
@@ -502,51 +484,4 @@ func (c *DockerCommand) execCaptured(ctx context.Context, in io.Reader, dir stri
 		return "", err
 	}
 	return out.String(), nil
-}
-
-func loadAuthFromConfig(conf *configfile.ConfigFile, registryHost string) (types.AuthConfig, error) {
-	return conf.AuthConfigs[registryHost], nil
-}
-
-func loadAuthFromCredentialsStore(ctx context.Context, credsStore string, registryHost string) (*CredentialHelperInput, error) {
-	var out strings.Builder
-	binary := DockerCredentialBinary(credsStore)
-	cmd := exec.CommandContext(ctx, binary, "get")
-	cmd.Env = os.Environ()
-	cmd.Stdout = &out
-	cmd.Stderr = &out
-	stdin, err := cmd.StdinPipe()
-	if err != nil {
-		return nil, err
-	}
-	defer stdin.Close()
-	console.Debug("$ " + strings.Join(cmd.Args, " "))
-	err = cmd.Start()
-	if err != nil {
-		return nil, err
-	}
-	_, err = io.WriteString(stdin, registryHost)
-	if err != nil {
-		return nil, err
-	}
-	err = stdin.Close()
-	if err != nil {
-		return nil, err
-	}
-	err = cmd.Wait()
-	if err != nil {
-		return nil, fmt.Errorf("exec wait error: %w", err)
-	}
-
-	var config CredentialHelperInput
-	err = json.Unmarshal([]byte(out.String()), &config)
-	if err != nil {
-		return nil, err
-	}
-
-	return &config, nil
-}
-
-func DockerCredentialBinary(credsStore string) string {
-	return "docker-credential-" + credsStore
 }
